+++ src/tests/test_benchmark.py
@@ -1,0 +1,22 @@
+import unittest
+
+from src.stac_utils.benchmark import make_msg, benchmark, Benchmark
+
+
+class TestBenchmark(unittest.TestCase):
+    def test_make_msg(self):
+        """ Test make message """
+
+    def test_benchmark_decorator(self):
+        """ Test benchmark decorator"""
+
+    def test_benchmark_class_init(self):
+        """ Test Benchmark class init """
+
+    def test_benchmark_class_context(self):
+        """ Test Benchmark class """
+
+
+
+if __name__ == '__main__':
+    unittest.main()
